+++ packages/react-article-components/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## UNRELEASED
 
+### 1.0.12
+
+- Add margin between topic and title of leading if there's no subtitle. By changing margin-top of subtitle to the margin-bottom of topic.
+- Change image diff images order
+
 ## RELEASE
 
 ### 1.0.11